src/favicon as defined in gatsby-config is not found
We added a preBootstrap lifecycle to ensure that a src/favicon.png
always exists. We chose the `onPreBootstrap` lifecycle because this is
a file that makes sense to give the user control over.

```
Error: icon (src/favicon.png) does not exist as defined in gatsby-config.js. Make sure the file exists relative to the root of the site.
```